refactor: change data type from float to decimal(30, 15)

+++ database/migrations/2025_06_03_145823_create_kriteria_bidang_industri.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('bidang_industri_id')->constrained('bidang_industri')->onDelete('cascade');
             $table->integer('rank');
-            $table->float('bobot');
+            $table->decimal('bobot', 30, 15);
 
             $table->timestamps();
         });
